First Attempt
Real time following the position of the user and draw that path on to the map with how far, how long have the user travelled with the statistical record of daily activities. Contains a foreground service to allow the user to control the App via the notification. The string file needs to be altered for your own google map key.